+++ zoomimage-core/src/androidMain/kotlin/com/github/panpf/zoomimage/util/AndroidLoggerPipeline.kt
@@ -25,14 +25,14 @@ actual fun createLogPipeline(): Logger.Pipeline = AndroidLogPipeline()
  */
 class AndroidLogPipeline : Logger.Pipeline {
 
-    override fun log(level: Int, tag: String, msg: String, tr: Throwable?) {
+    override fun log(level: Logger.Level, tag: String, msg: String, tr: Throwable?) {
         when (level) {
-            Logger.VERBOSE -> Log.v(tag, msg, tr)
-            Logger.DEBUG -> Log.d(tag, msg, tr)
-            Logger.INFO -> Log.i(tag, msg, tr)
-            Logger.WARN -> Log.w(tag, msg, tr)
-            Logger.ERROR -> Log.e(tag, msg, tr)
-            Logger.ASSERT -> Log.wtf(tag, msg, tr)
+            Logger.Level.Verbose -> Log.v(tag, msg, tr)
+            Logger.Level.Debug -> Log.d(tag, msg, tr)
+            Logger.Level.Info -> Log.i(tag, msg, tr)
+            Logger.Level.Warn -> Log.w(tag, msg, tr)
+            Logger.Level.Error -> Log.e(tag, msg, tr)
+            Logger.Level.Assert -> Log.wtf(tag, msg, tr)
         }
     }