+++ zoomimage-core/src/nonAndroidMain/kotlin/com/github/panpf/zoomimage/util/Logger.nonAndroid.kt
@@ -23,12 +23,12 @@ actual fun createLogPipeline(): Logger.Pipeline = PrintlnLogPipeline()
  */
 class PrintlnLogPipeline : Logger.Pipeline {
 
-    override fun log(level: Int, tag: String, msg: String, tr: Throwable?) {
+    override fun log(level: Logger.Level, tag: String, msg: String, tr: Throwable?) {
         if (tr != null) {
             val trString = tr.stackTraceToString()
-            println("${Logger.levelName(level)}. $tag. $msg. \n$trString")
+            println("${level}. $tag. $msg. \n$trString")
         } else {
-            println("${Logger.levelName(level)}. $tag. $msg")
+            println("${level}. $tag. $msg")
         }
     }