+++ zoomimage-view/src/main/kotlin/com/github/panpf/zoomimage/view/subsampling/SubsamplingEngine.kt
@@ -396,7 +396,7 @@ class SubsamplingEngine constructor(
             } else {
                 val exception = result.exceptionOrNull()!! as CreateTileDecoderException
                 this@SubsamplingEngine._imageInfoState.value = exception.imageInfo
-                val level = if (exception.skipped) Logger.DEBUG else Logger.ERROR
+                val level = if (exception.skipped) Logger.Level.Debug else Logger.Level.Error
                 val type = if (exception.skipped) "skipped" else "error"
                 logger.log(level) {
                     "resetTileDecoder:$caller. $type, ${exception.message}. " +